Add files via upload
cleaned out our original individual files, just leaving the "combined" file we are working from.

to test, download and set up MySQL on your computer and import the two tables I emailed a while back (they are empty - just column setup).  

All the pages will load, but not all of them are actually talking to the database.  

Here are some bugs/issues to work on so far:
1.  Registration:  when you click to register from the Login page, it doesn't verify your email is an email address - add check?
2.  Profile page (added from Madison's sample code - thank you!):  right now, only links from the login page.  We could probably clean up the navigation on the whole - have a navbar or menu drop down at the top, rather than buttons at the bottom for everything.  Edit password function gives confirmation message, but isn't actually changing the user table in the database.
3.  Add/Edit Inventory:  isn't adding items to the table yet, also the image code on the code behind file (lines 39-44) was giving an error so I just commented out for now to make sure the rest would load
4.  ItemDB code:  the Select method should link to the userID, so that different logins can see only their own inventory, but I'm getting an error when I add the "WHERE" clause.